Restructuring the frontend routes folder and renaming some files
