+++ client/src/routes/auth/classic/TasksRoute.jsx
@@ -1,15 +1,10 @@
-import DashboardPage from "../../pages/auth/dashboard/classic/Dashboard";
-import WrongRoute from "../public/WrongRoute";
-
 import { useLocation, useNavigate } from "react-router";
+import TasksPage from "../../../pages/auth/tasks/Tasks";
 import { useEffect, useState } from "react";
+import WrongRoute from "../../public/WrongRoute";
 import bcrypt from "bcryptjs";
 
-export default function DashboardRouteClassic({
-  isAuthenticated,
-  setAuthentication,
-  setPreviewModernUI,
-}) {
+export default function TasksRoute({ isAuthenticated, setAuthentication, setPreviewModernUI }) {
   const [session, setSession] = useState("");
 
   let navigate = useNavigate();
@@ -64,15 +59,15 @@ export default function DashboardRouteClassic({
       if (userLoggedOut) {
         navigate("/signin");
       } else {
-        navigate("/signin?redirect=/auth/user/dashboard");
+        navigate("/signin?redirect=/auth/user/classic/tasks");
       }
     }
-  }, []);
+  });
 
-  if ((isAuthenticated || userAuthenticated) && !userLoggedOut) {
+  if (isAuthenticated || userAuthenticated) {
     if (username === userAuthenticated.user) {
       return (
-        <DashboardPage
+        <TasksPage
           user={userAuthenticated.user}
           userId={userAuthenticated.userId}
           setAuthentication={setAuthentication}
